refactor(question-set): remove redundant
